debian: Move conf files to /usr/lib
debhelper assumes everything in /etc is a conf file.
This means they won't be removed on removal (only on purge)
and they won't be upgraded/installed if edited/removed manually.

For example, this is a problem if user wants to swap their installation from
phosh to cutie. In this scenario, the phosh conf for
`android-service@hwcomposer.service` still remains on the system and
prevents cutie from starting: "Could not find phosh.service"

Therefore, we should install these files in /usr/lib
(systemd still reads from there).